Use dataloaders.
