+++ server/package.json
@@ -14,6 +14,7 @@
     "apollo-server-express": "^2.14.1",
     "axios": "^0.19.2",
     "bcryptjs": "^2.4.3",
+    "dataloader": "^2.0.0",
     "graphql": "^15.0.0",
     "jsonwebtoken": "^8.5.1",
     "merge-graphql-schemas": "^1.7.8",
