+++ server/src/db/queries/progress/updateStepStatus.js
