+++ server/src/graphql/resolvers/pathway.js
@@ -2,7 +2,7 @@ const queries = require('../../db/queries/queries')
 
 const resolver = {
     Mutation: {
-        async createUpdatePathway(_, {id, name, steps, tags}) {
+        async createUpdatePathway(_, {id, name, steps}) {
             const query = queries.pathway.createUpdatePathway({
                 id, name, steps
             })
@@ -50,47 +50,6 @@ const resolver = {
                 return { status: 'ERROR', message: e.toString() }
             }
         },
-        async updateUserStepStatus(_, {username, stepId, status}) {
-            const query = queries.user.updateStepStatus({
-                username, stepId, status
-            })
-
-            try {
-                await query.run()
-                return { status: 'OK', message: null}
-            } catch (e) {
-                return {status: 'ERROR', message: e.toString() }
-            }
-        }
-    },
-    Step: {
-        async userStatus(parent, _, context) {
-            const user = context.user
-            if (user == null) {
-                return "NOT_STARTED"
-            }
-
-            const username = user.username
-            const stepId = parent.id
-            const query = queries.user.getStepStatus({
-                username, stepId
-            })
-
-            let status = ''
-
-            const result = await query.run()
-
-            if (result.records.length == 0) {
-                status = 'NOT_STARTED'
-            } else {
-                status = result.records[0].get(0)
-            }
-
-
-            console.log(result)
-
-            return status
-        }
     }
 }
